requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
airtable-python-wrapper==0.15.3
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
discord.py==2.3.2
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
multidict==6.0.5
notion-client==2.2.1
openai==1.14.2
pydantic==2.6.4
pydantic_core==2.16.3
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
